+++ src/jvmMain/kotlin/com/sunnychung/application/multiplatform/giantlogviewer/annotation/Persisted.kt
@@ -1,0 +1,3 @@
+package com.sunnychung.application.multiplatform.giantlogviewer.annotation
+
+annotation class Persisted()