+++ src/jvmMain/kotlin/com/sunnychung/application/multiplatform/giantlogviewer/ux/TextToggleButton.kt
@@ -2,6 +2,7 @@ package com.sunnychung.application.multiplatform.giantlogviewer.ux
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.text.BasicText
 import androidx.compose.runtime.Composable
@@ -12,6 +13,7 @@ import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.sunnychung.application.multiplatform.giantlogviewer.ux.local.LocalColor
 import com.sunnychung.application.multiplatform.giantlogviewer.ux.local.LocalFont
 import com.sunnychung.lib.multiplatform.bigtext.extension.runIf
 
@@ -22,22 +24,31 @@ fun TextToggleButton(
     fontFamily: FontFamily = LocalFont.current.normalFontFamily,
     isSelected: Boolean,
     isEnabled: Boolean = true,
+    innerPadding: PaddingValues = PaddingValues(horizontal = 2.dp, vertical = 4.dp),
     onToggle: (Boolean) -> Unit,
 ) {
-//    val colours = LocalColor.current
+    val colors = LocalColor.current
     BasicText(
         text = text,
         style = TextStyle(
             fontFamily = fontFamily,
-            color = if (isEnabled) Color.White else Color(.6f, .6f, .6f),
+            color = if (isEnabled) {
+                if (isSelected) {
+                    colors.toggleButtonOnText
+                } else {
+                    colors.toggleButtonOffText
+                }
+            } else {
+                colors.toggleButtonDisabledText
+            },
             fontSize = 14.sp,
             textAlign = TextAlign.Center,
         ),
         modifier = modifier
-            .background(if (isSelected && isEnabled) Color(.2f, .2f, .7f) else Color(0f, 0f, .3f))
+            .background(if (isSelected && isEnabled) colors.toggleButtonOnBackground else colors.toggleButtonOffBackground)
             .runIf(isEnabled) {
                 clickable { onToggle(!isSelected) }
             }
-            .padding(horizontal = 2.dp, vertical = 4.dp)
+            .padding(innerPadding)
     )
 }