+++ src/jvmMain/kotlin/com/sunnychung/application/multiplatform/giantlogviewer/ux/local/AppColor.kt
@@ -1,0 +1,6 @@
+package com.sunnychung.application.multiplatform.giantlogviewer.ux.local
+
+import androidx.compose.runtime.compositionLocalOf
+import com.sunnychung.application.multiplatform.giantlogviewer.document.lightColorTheme
+
+internal val LocalColor = compositionLocalOf { lightColorTheme() }